Multitude of fixes
Removed shareNumber, triviaNumber, and learnNumber variables. These are redundant now that we are reshuffling after every trial. We now instead just reference the first two items in the dataframe (0 and 1)

RemoveUsedItem() was not giving new indexes after dropping an item, so the indexes would go from 0, 1, 3, 5 etc. This is why after 2 trials it would error out. Fixed